moving the log filter function to hubblestack/log.py from conf_publisher
