+++ hubblestack/extmods/modules/conf_publisher.py
@@ -41,7 +41,7 @@ def publish(report_directly_to_splunk=True, remove_dots=True, *args):
             if arg in  __opts__:
                 opts_to_log[arg] = __opts__[arg]
 
-    filtered_conf = _filter_config(opts_to_log, remove_dots=remove_dots)
+    filtered_conf = hubblestack.log.filter_logs(opts_to_log, remove_dots=remove_dots)
 
     if report_directly_to_splunk:
         hubblestack.log.emit_to_splunk(filtered_conf, 'INFO', 'hubblestack.hubble_config')
@@ -49,30 +49,3 @@ def publish(report_directly_to_splunk=True, remove_dots=True, *args):
 
     return filtered_conf
 
-
-def _filter_config(opts_to_log, remove_dots=True):
-    '''
-    Filters out keys containing certain patterns to avoid sensitive information being sent to splunk
-    '''
-    patterns_to_filter = ["password", "token", "passphrase", "privkey", "keyid", "s3.key"]
-    filtered_conf = _remove_sensitive_info(opts_to_log, patterns_to_filter)
-    if remove_dots:
-        for key in filtered_conf.keys():
-            if '.' in key:
-                filtered_conf[key.replace('.', '_')] = filtered_conf.pop(key)
-    return filtered_conf
-
-
-def _remove_sensitive_info(obj, patterns_to_filter):
-    '''
-    Filter known sensitive info
-    '''
-    if isinstance(obj, dict):
-         obj = {
-             key: _remove_sensitive_info(value, patterns_to_filter)
-             for key, value in obj.iteritems()
-             if not any(patt in key for patt in patterns_to_filter)}
-    elif isinstance(obj, list):
-         obj = [_remove_sensitive_info(item, patterns_to_filter)
-                    for item in obj]
-    return obj
