+++ hubblestack/log.py
@@ -7,8 +7,12 @@ from __future__ import print_function
 
 import logging
 import time
+import copy
 
 import hubblestack.splunklogging
+
+# These patterns will not be logged by "conf_publisher" module and emit_to_splunk
+PATTERNS_TO_FILTER = ["password", "token", "passphrase", "privkey", "keyid", "s3.key"]
 
 # While hubble doesn't use these, salt modules can, so let's define them anyway
 SPLUNK = logging.SPLUNK = 25
@@ -154,10 +158,17 @@ def setup_splunk_logger():
     SPLUNK_HANDLER = handler
 
 
-def emit_to_splunk(message, level, name):
+def emit_to_splunk(message, level, name, filter_logs=False):
     '''
     Emit a single message to splunk
     '''
+
+    if filter_logs:
+        # copying the original message so that it does not
+        # get modified while filtering
+        copy_of_message = copy.deepcopy(message)
+        message = filter_logs(copy_of_message, remove_dots=True)
+
     if SPLUNK_HANDLER is None:
         return False
     handler = SPLUNK_HANDLER
@@ -177,3 +188,31 @@ def workaround_salt_log_handler_queues():
     sls.LOGGING_NULL_HANDLER.sync_with_handlers([flh])
     # if flh.count > 0:
     #     log.info("pretended to handle %d logging record(s) for salt.log.setup.LOGGING_*_HANDLER", flh.count)
+
+def filter_logs(opts_to_log, remove_dots=True):
+    '''
+    Filters out keys containing certain patterns to avoid sensitive information being sent to logs
+    Works on dictionaries and lists
+    This function was located at extmods/modules/conf_publisher.py previously
+    '''
+    filtered_conf = _remove_sensitive_info(opts_to_log, PATTERNS_TO_FILTER)
+    if remove_dots:
+        for key in filtered_conf.keys():
+            if '.' in key:
+                filtered_conf[key.replace('.', '_')] = filtered_conf.pop(key)
+    return filtered_conf
+
+
+def _remove_sensitive_info(obj, patterns_to_filter):
+    '''
+    Filter known sensitive info
+    '''
+    if isinstance(obj, dict):
+         obj = {
+             key: _remove_sensitive_info(value, patterns_to_filter)
+             for key, value in obj.iteritems()
+             if not any(patt in key for patt in patterns_to_filter)}
+    elif isinstance(obj, list):
+         obj = [_remove_sensitive_info(item, patterns_to_filter)
+                    for item in obj]
+    return obj
